Housework/Final/requirement.txt
django
djangorestframework
djangorestframework-simplejwt
openai